wip(poly sync): extract AST parser to separate component
