+++ components/polylith/libs/__init__.py
@@ -1,13 +1,4 @@
 from polylith.libs import report
-from polylith.libs.grouping import (
-    extract_third_party_imports,
-    fetch_all_imports,
-    get_third_party_imports,
-)
+from polylith.libs.grouping import extract_third_party_imports, get_third_party_imports
 
-__all__ = [
-    "report",
-    "extract_third_party_imports",
-    "fetch_all_imports",
-    "get_third_party_imports",
-]
+__all__ = ["report", "extract_third_party_imports", "get_third_party_imports"]
